+++ printer/ast_printer/ast_printer.go
@@ -48,6 +48,8 @@ func PrintRecursive(node ast.Node[ast.NodeType], sb *strings.Builder, depth int)
 		return printBinaryOperation(node, sb, depth)
 	case *ast.StateVariableDeclaration:
 		return printStateVariableDeclaration(node, sb, depth)
+	case *ast.Emit:
+		return printEmit(node, sb, depth)
 	default:
 		if node.GetType() == ast_pb.NodeType_SOURCE_UNIT {
 			return printSourceUnit(node, sb, depth)
@@ -101,4 +103,49 @@ func writeStrings(sb *strings.Builder, s ...string) {
 
 func indentString(s string, depth int) string {
 	return strings.Repeat(" ", depth*INDENT_SIZE) + s
+}
+
+func getStorageLocationString(storage ast_pb.StorageLocation) string {
+	switch storage {
+	case ast_pb.StorageLocation_DEFAULT:
+		return ""
+	case ast_pb.StorageLocation_MEMORY:
+		return "memory"
+	case ast_pb.StorageLocation_STORAGE:
+		return "storage"
+	case ast_pb.StorageLocation_CALLDATA:
+		return "calldata"
+	default:
+		return ""
+	}
+}
+
+func getVisibilityString(visibility ast_pb.Visibility) string {
+	switch visibility {
+	case ast_pb.Visibility_INTERNAL:
+		return "internal"
+	case ast_pb.Visibility_PUBLIC:
+		return "public"
+	case ast_pb.Visibility_EXTERNAL:
+		return "external"
+	case ast_pb.Visibility_PRIVATE:
+		return "private"
+	default:
+		return ""
+	}
+}
+
+func getStateMutabilityString(mut ast_pb.Mutability) string {
+	switch mut {
+	case ast_pb.Mutability_PURE:
+		return "pure"
+	case ast_pb.Mutability_VIEW:
+		return "view"
+	case ast_pb.Mutability_NONPAYABLE:
+		return ""
+	case ast_pb.Mutability_PAYABLE:
+		return "payable"
+	default:
+		return ""
+	}
 }
